Handle KeyError Exception Gracefully
All embys may not report the same media item data back. Update to handle missing data items from emby. Script uses this data to build detailed output about what is watched and what will and will not be deleted.
Now if any of the extra data is missing we fall back to the basic data needed to delete the episode and continue on to the next media item.