Config update and Adding Audio
Breaking config update. Will require re-creating config file.
Update config file to explain what each variable in the config file does.
Start the addition of audio as a media. As of now only favorite audio tracks are preserved.

Future release will have ability to delete or not delete based on favorite track, favorite artist, favorite album.
Favorite artist will have the ability to "keep track when favorite artist is the primary artist of a track or album" or "keep track when favorite artist is the primary or featured artist of a track or album". This of course is dependent on the data Emby/Jellyfin return.